+++ src/sass/_weather.scss
@@ -1,0 +1,12 @@
+.weather__card{
+
+}
+
+.weather__card__container{
+    background-color: yellow;
+}
+
+
+.weather__card li{
+    display: flex;
+}
